studs_service/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.9.0
asyncpg==0.30.0
billiard==4.2.1
celery==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
db==0.1.1
python-dotenv==1.1.0
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
kombu==5.5.3
numpy==1.24.3
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.4
python-dateutil==2.9.0.post0
redis==5.0.0
requests==2.32.3
scipy==1.15.2
simpleeval==1.0.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
